+++ airflow/providers/common/sql/CHANGELOG.rst
@@ -32,7 +32,9 @@ This release fixes a few errors that were introduced in common.sql operator whil
 * ``_process_output`` method in ``SQLExecuteQueryOperator`` has now consistent semantics and typing, it
   can also modify the returned (and stored in XCom) values in the operators that derive from the
   ``SQLExecuteQueryOperator``).
-* last description of the cursor whether to return scalar values are now stored in DBApiHook
+* descriptions of all returned results are stored as descriptions property in the DBApiHook
+* last description of the cursor whether to return single query results values are now exposed in
+  DBApiHook via last_description property.
 
 Lack of consistency in the operator caused ``1.3.0`` to be yanked - the ``1.3.0`` should not be used - if
 you have ``1.3.0`` installed, upgrade to ``1.3.1``.